Changed solution, changeDrive->changeDrives to indicate that more than one drive can be checked. Solution is still targeting SSD/NVME.
checkDerives.{services | timer} are for creating a Systemd service.
checkDrives_template.cfg is a template for the config file.

models holds supported models. Ideally, should be online/dynamic, but
for now hardcoded and embedded in solution.

identifyDrives.sh can be used to identify the drives this instance
can work with,, and form the base to the config file.

Makefile, to 'install', 'uninstall' the solution and service.

The solution now reads from the /etc/default/checkDrives.cfg for its
parameters. These can be overrriden from command line arguments.